Hiiiii

+++ SalesAndInentoryWeb_Application/SalesAndInentoryWeb_Application/ViewModel/PartyDetailsCreditNote.cs
@@ -22,7 +22,7 @@ namespace SalesAndInentoryWeb_Application.ViewModel
         public string VehicleNumber { get; set; }
         public DateTime DeliveryDate { get; set; }
         public string Barcode { get; set; }              
-        public int TaxAmount1 { get; set; }
+        public float TaxAmount1 { get; set; }
         public double Received { get; set; }
         public double RemainingBal { get; set; }
         public double CalTotal { get; set; }
